Add files via upload
Fixing mistake of marking people who had a recordeddate but sawlastpage!=1 (because they had said that they got the link not from UT and then were kicked out of the survey) as "having completed the survey" and thus not requiring a reminder. Also expanded date window for 'need reminder' to be -1 to 14 instead of 0 to 14, and added in a few stopifnot tests.